address diff comments
